+++ src/main/java/httpServer/routes/Router.java
@@ -52,15 +52,15 @@ public class Router {
 
     private HTTPResponseBuilder setBody(HTTPRequest httpRequest, HTTPResponseBuilder responseBuilder) {
         switch (httpRequest.resource) {
-            case "/echo_body":
-                responseBuilder = responseBuilder.setBody(httpRequest.body);
-                break;
-            case "/json_response":
-                responseBuilder = responseBuilder.setBody("{\"key1\":\"value1\",\"key2\":\"value2\"}");
-                break;
-            default:
-                responseBuilder = responseBuilder.setBody("");
-                break;
+          case "/echo_body":
+              responseBuilder = responseBuilder.setBody(httpRequest.body);
+              break;
+          case "/json_response":
+              responseBuilder = responseBuilder.setBody("{\"key1\":\"value1\",\"key2\":\"value2\"}");
+              break;
+          default:
+              responseBuilder = responseBuilder.setBody("");
+              break;
         }
         return responseBuilder;
     }
